Better child category support.
Now loading the category inheritance tree completely (instead of just
one level as previously). This is implemented with recursive query to
categoryinheritance table.

When listing bookmarks changed to only display root categories in the
selection form.

When querying for bookmarks in a category changed to also query in all
the child categories.