Fix loading of bookmarks in multiple categories.
- When multiple categories where selected in bookmark view the query was
actually "category1 OR category2 OR ...". Fixed this to return the
correct "category1 AND category2 AND ..." so only bookmarks that are in
all the selected categories are displayed.

- If only a single category is selected, then the children of the
category are loaded, and the bookmarks in the category are queried with
"category OR child_category1 OR child_category2" so all the bookmarks in
the category and child categories are shown.